requirements.txt
pandas==2.2.0
websocket-client==1.7.0
pyarrow==15.0.0
matplotlib==3.8.2
requests==2.31.0
websockets==12.0
python-dotenv==1.0.1
asyncpg==0.29.0
aiohttp==3.9.1